refactor: Add ShortFormPlayer components
