+++ components/ShortFormPlayer/Video/EmbedPlayer/Mask.tsx
